Update download script with better case-handling

+++ utils/get_input.py
@@ -2,13 +2,8 @@
 # -*- coding: utf-8 -*-
 
 """
-Utility scritp that works with crontab to fetch the input puzzle for the current year
-
-Crontab config:
-$  crontab -e
-$  15 10 1-25 12 * python3 $PATH/advent-of-code/utils/get_input.py > /dev/null 2>&1
-
-## UPDATE: cronjob deprecated...##
+Use this script to download the advent-of-code input file automatically
+It requires an hidden .cookie file with the cookies for your advent-of-code session
 """
 
 __author__ = 'Guido Minieri'
@@ -16,48 +11,61 @@ __license__ = 'GPL'
 
 
 from datetime import date
-from time import sleep
 import os
 import requests
+import sys
 
-
-# wait for internet connection to be restored
-sleep(3)
-
-# get the directory of the file
+# get the root directory
 abspath = os.path.abspath(__file__)
 dname = os.path.dirname(abspath)
 os.chdir(dname)
 
-
+# fetch today's date and assign year and day
 today = date.today()
-
 year = today.year
+month = today.month
 day = today.day
 
+# exit the program if there are no challenges today
+if month != 12 or day > 25:
+    print("There are no challanges today!")
+    sys.exit()
+
+# go back to advent-of-code repository
+os.chdir('./..')
+
+# move into year folder and create it if it's the first year
+if os.path.isdir(f'./{year}'):
+    os.chdir(f'./{year}')
+else:
+    os.mkdir(f'./{year}')
+    os.chdir(f'./{year}')
+
+# check if folder exist then create it or move into it (zfill for padding in file name)
+foldername = str(day).zfill(2)
+if os.path.isdir(f'./{foldername}'):
+    os.chdir(f'./{foldername}')
+else:
+    os.mkdir(f'./{foldername}')
+    os.chdir(f'./{foldername}')
+
+# if file exists exit
+if os.path.exists('./input.txt'):
+    print('Input file has already been downloaded')
+    print('Happy Coding!')
+    sys.exit()
+
+# get cookies
 with open('.cookie', 'r') as f:
     cookie = f.read().strip()
-
-cookies = {'session': cookie}  # define cookie for the session
-url = f"https://adventofcode.com/{year}/day/{day}/input"
+cookies = {'session': cookie}
 
 # fetch data
+url = f"https://adventofcode.com/{year}/day/{day}/input"
 r = requests.get(url, cookies=cookies)
 res = r.text
 
-os.chdir('./..')  # go back to root dir
-
-# handle year
-if not os.path.isdir(f'./{year}'):  # if first day of december
-    os.mkdir(f'./{year}')
-    os.chdir(f'./{year}')
-else:
-    os.chdir(f'./{year}')
-
-
-# zfill for padding
-os.mkdir(f'./{str(day).zfill(2)}')
-os.chdir(f'./{str(day).zfill(2)}')
-
+# write the input puzzle in the input file
 with open('input.txt', 'w') as f:
     f.write(res)
+print(f'Data fetched: input.txt written in {year}/{day}')
